Pin Guava version to 20.0
== DETAILS

Dependencies in released products should probably never have an
open-ended dependency as this one had on Guava. It will resolve to the
latest version of an artifact which, following semantic versioning,
may break the build. That is the case with this library now if Guava
resolves to 23.0.

So this commit pins the version to 20.0.