Update to Guava 23
== DETAILS

Updating to Guava 23 requires updating the target compatibility to
Java 8. Otherwise, the changes are minor.

There is a small change in behavior. When a Retryer is built with a
TimeLimiter, timeout causes an ExecutionException to be thrown. In
previous versions, the cause of that exception was an instance of
com.google.common.util.concurrent.UncheckedTimeoutException.  Now it
is an instance of java.util.concurrent.TimeoutException.