Remove two tests that fail in Intellij
The tests check what happens when null is passed as a parameter to the
method in question. The method has a Guava Precondition to check for
non-null. However, when Intellij runs the tests they fail with an
IllegalArgumentException. Apparently something in Intellij is
intercepting the method call and performaning its own check for
non-null.

I'm just removing the tests: having all tests pass in Intellij is way
more important than testing what happens when you pass null to a
method that is documented to require a non-null parmeter.