Refactor AttemptTimeLimiter: remove class-level parameter
The AttemptTimeLimiter had a class level parameter. It was used to
specify the type of the Callable AttemptTimeLimiter.call takes.

However to deal with issue #13, we will need to limit time for
Runnables, which do not have generic parameters. My plan is to simply
reuse AttemptTimeLimiter, but add a class. Before I do that, though I
need to remove the class level parameter.

So I removeed the class-level parameter from AttemptTimeLimiter.
AttemptTimeLimiter.call now has a method-level generic parameter so it
can return the correct type.