Revert "Add Retryer methods callUnchecked and runUnchecked"
This reverts commit db6390c128b000fa17df8462a9be0ac3eacdff7e.

After considering how to deal with InterruptedExceptions (which can
happen because of the Thread.sleep method between attempts), I think
the unchecked methods are probably a bad idea.

The original guava-retrying library always wrapped the
InterruptedExceptions in a RetryException, and reset the interrupted
flag. However, I think that's probably not a great idea either:
callers are unlikely to check that flag and finish up their threads
immediately.

So I am going to remove the uncheck methods. Then I am going to cause
the call and run methods propagate InterruptedExceptions.