Remove unchecked methods
== DETAILS

This removes the callUnchecked and runUnchecked methods. Considering
that callers need to deal with the InterruptedException that can be
thrown by the sleep between attempts, I'm just propagating the
InterruptedException. My attempt to have an method that could take a
lambada converted to a Runnable with no exception handling was
laudable, but misguided.

I also changed the behavior so the call and run methods never just
propagate the exception from the Callable or Runnable; they always use
a RetryException.